[FIX] commit anterior

+++ WebAPI-My-Home-Library/Models/UsuarioLivro.cs
@@ -12,7 +12,7 @@ namespace WebAPI_My_Home_Library.Models
     {
         [Key]
         [Column("ide_usuario_livro")]
-        public int Ide_Usuario_Livro { get; set; }
+        public long Ide_Usuario_Livro { get; set; }
 
         [Column("ide_usuario")]
         public long Ide_Usuario { get; set; }
